ai-agent/requirements.txt
# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# API Clients
requests>=2.31.0
httpx>=0.27.0

# Tavily API
tavily-python>=0.3.0

# LLM Integration
groq>=0.4.0

# Data Processing (using conda-forge or pre-built wheels)
pandas>=2.2.0
numpy>=1.26.0

# Utilities
pydantic>=2.6.0
python-dateutil>=2.8.0
